Fix all PR review comments from performance-optimizations branch
P1 MANDATORY fixes:
1. Convert go.mod replace() block to single-line format (fixes Docker build compatibility)
2. Keep local deps paths (./deps/) as they are used for local development
3. Restore phpext.go cgo CFLAGS to -I../frankenphp to match Docker build flow

Copilot/Security fixes:
4. Add pool size validation (size > 0) in NewPool
5. Add two-value receive check in Pool.Get to detect closed channels
6. Add closed-check in Pool.Put with error return on channel closure
7. Implement on-demand instantiation in Registry.Get for lazy registries
8. Change argsLen parameter from C.int to C.size_t in go_wasm_call with bounds check
9. Change length cast from (int) to (size_t) in wasmplugin.c line 239

Additional improvements:
- Add ErrPoolClosed sentinel error for pool closure detection
- Add Reset() method to Registry for future per-request cleanup
- Add lockFor() helper for per-plugin mutual exclusion in lazy registries
- Improve comments and documentation for clarity

+++ phpext/wasmplugin.c
@@ -236,7 +236,7 @@ PHP_METHOD(Wasm, call)
         ZSTR_VAL(intern->name),
         ZSTR_VAL(function_name),
         ZSTR_VAL(args_json.s),
-        (int)ZSTR_LEN(args_json.s)
+        (size_t)ZSTR_LEN(args_json.s)
     );
 
     smart_str_free(&args_json);
